pyproject.toml
[project]
name = "video-transcribe"
version = "0.1.0"
description = "Automated video transcription with speaker diarization and meeting summaries"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "v" }
]

dependencies = [
    "ffmpeg-python>=0.2.0",
    "pydub>=0.25.1",
    "click>=8.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
video-transcribe = "video_transcribe.__main__:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
